Update openfb.js to the new facebook api
Hi, can you upgrade this file with the new 2.0 version of facebook login, because the 1.0 version will expire soon. 

"With the launch of v2.0 of the API, we have two versions in service at the same time:

Version 1.0, which is what we call the API as it existed the day before v2.0 was launched. We'll support v1.0 for one year and it will expire on April 30th, 2015.
Version 2.0, which is what this upgrade guide covers. Version 2.0 is supported for at least two years. At the earliest, it will expire on April 30th, 2016."
you can replace this https://www.facebook.com/dialog/oauth? by yhis https://www.facebook.com/v2.0/dialog/oauth?